Need to leave urgently. Need to do: 1) HealthBar to UnitView 2) Move view logic from BaseUnit to UnitView 3) Rename BaseUnit to BaseUnitController 4) Fix BaseMode init values 5) And so on...

+++ VIKING-Raiders-AFK/Assets/_Scripts/UnitScripts/HealthBar.cs
@@ -1,4 +1,3 @@
-using UnitScripts;
 using UnityEngine;
 using UnityEngine.UI;
 
@@ -7,14 +6,14 @@ public class HealthBar : MonoBehaviour
     [SerializeField] private Slider slider;
     [SerializeField] private Gradient gradient;
     [SerializeField] private Image fill;
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         slider.maxValue = health;
         slider.value = health;
         fill.color = gradient.Evaluate(1f);
     }
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
         fill.color = gradient.Evaluate(slider.normalizedValue);
